// client/models/workPosition/detail.model.ts
export interface DetailModel {
    companyDescription: string;
    description: string;
    requesting: string;
    offering: string;
    customField1Name: string;
    customField2Name: string;
    customField3Name: string;
    customField1Value: string;
    customField2Value: string;
    customField3Value: string;
}